requirements.txt
boto==2.41.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-crispy-forms==1.6.0
django-postgrespool==0.3.0
django-storages==1.4.1
django-toolbelt==0.0.1
gunicorn==19.3.0
Pillow==3.2.0
psycopg2==2.6
sorl-thumbnail==12.3
SQLAlchemy==1.0.4
static3==0.6.1
whitenoise==1.0.6
